# Cache actor activations