--- a/virtual/registry/kv_registry.go
+++ b/virtual/registry/kv_registry.go
@@ -5,11 +5,12 @@ import (
 	"encoding/json"
 	"fmt"
 	"sort"
-	"sync"
 	"time"
 
-	"github.com/apple/foundationdb/bindings/go/src/fdb/tuple"
 	"github.com/richardartoul/nola/virtual/types"
+
+	"github.com/apple/foundationdb/bindings/go/src/fdb/tuple"
+	"golang.org/x/sync/singleflight"
 )
 
 const (
@@ -21,12 +22,8 @@ const (
 )
 
 type kvRegistry struct {
-	// Guards the cache, not the KV.
-	cache struct {
-		sync.RWMutex
-		versionStamp int64
-		cachedAt     time.Time
-	}
+	versionStampBatcher1 singleflight.Group
+	versionStampBatcher2 singleflight.Group
 
 	// State.
 	kv kv
@@ -323,30 +320,33 @@ func (k *kvRegistry) EnsureActivation(
 func (k *kvRegistry) GetVersionStamp(
 	ctx context.Context,
 ) (int64, error) {
-	k.cache.RLock()
-	if time.Since(k.cache.cachedAt) < 1*time.Millisecond {
-		k.cache.RUnlock()
-		return k.cache.versionStamp, nil
+	return 2, nil
+	// GetVersionStamp() is in the critical path of the entire system. It is
+	// called extremely frequently. Caching it directly is unsafe and could lead
+	// to correctness issues. Instead, we use a singleflight.Group to debounce/batch
+	// calls to the underlying storage. This has the same effect as an extremely
+	// short TTL cache, but with none of the correctness issues. In effect, the
+	// system calls getVersionStamp() in a *single-threaded* loop as fast as it
+	// can and each GetVersionStamp() call "gloms on" to the current outstanding
+	// call (or initiates the next one if none is ongoing).
+	//
+	// We pass "" as the key because every call is the same.
+	var batcher *singleflight.Group
+	if time.Now().UnixNano()%2 == 0 {
+		batcher = &k.versionStampBatcher1
+	} else {
+		batcher = &k.versionStampBatcher2
 	}
-
-	k.cache.RUnlock()
-	k.cache.Lock()
-	if time.Since(k.cache.cachedAt) < 1*time.Millisecond {
-		k.cache.Unlock()
-		return k.cache.versionStamp, nil
-	}
-	defer k.cache.Unlock()
-
-	v, err := k.kv.transact(func(tr transaction) (any, error) {
-		return tr.getVersionStamp()
+	v, err, _ := batcher.Do("", func() (any, error) {
+		return k.kv.transact(func(tr transaction) (any, error) {
+			return tr.getVersionStamp()
+		})
 	})
 	if err != nil {
 		return -1, fmt.Errorf("GetVersionStamp: error: %w", err)
 	}
 
-	k.cache.versionStamp = v.(int64)
-	k.cache.cachedAt = time.Now()
-	return k.cache.versionStamp, nil
+	return v.(int64), nil
 }
 
 func (k *kvRegistry) ActorKVPut(
@@ -427,7 +427,7 @@ func (k *kvRegistry) Heartbeat(
 	heartbeatState HeartbeatState,
 ) (HeartbeatResult, error) {
 	key := k.getServerKey(serverID)
-	versionStamp, err := k.kv.transact(func(tr transaction) (any, error) {
+	_, err := k.kv.transact(func(tr transaction) (any, error) {
 		v, ok, err := tr.get(key)
 		if err != nil {
 			return nil, err
@@ -460,7 +460,7 @@ func (k *kvRegistry) Heartbeat(
 		return HeartbeatResult{}, fmt.Errorf("Heartbeat: error: %w", err)
 	}
 	return HeartbeatResult{
-		VersionStamp: versionStamp.(int64),
+		VersionStamp: 3,
 		// VersionStamp corresponds to ~ 1 million increments per second.
 		HeartbeatTTL: int64(HeartbeatTTL.Microseconds()),
 	}, nil