@aratz-lasa 2023-01-05
Amazing redesign! 

What command did you run for the benchmark? I would like to replicate it locally for playing around with it.