@richardartoul 2023-01-06
@aratz-lasa 

```
/opt/homebrew/bin/go test -timeout 60s -run ^TestBenchmarkFoundationRegistryInvoke$ github.com/richardartoul/nola/virtual
```